[Refactory] 카카오 JSON 역직렬화 로직 수정 #15
- 패키지 변경
- @JvmStatic을 사용하여 JSON 역직렬화 로직 수정
