+++ src/main/kotlin/org/washcode/washpang/controller/CommPageController.kt
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Controller
 import org.springframework.ui.Model
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RequestParam
-import org.washcode.washpang.global.client.KakaoClient
+import org.washcode.washpang.global.comm.client.KakaoClient
 
 
 @Controller
